rabbitmq/requirements.txt
coloredlogs==15.0.1
humanfriendly==10.0
pika==1.3.2
pillow==10.4.0
python-dotenv==1.0.1
